Implemented reference assembly discovery for the function invocation part.
