+++ src/Kubeless.Core/Models/MethodInvoker.cs
@@ -1,0 +1,4 @@
+namespace Kubeless.Core.Models
+{
+    public delegate object MethodInvoker(params object[] parameters);
+}
